refactoring: move parsing and allies code out of config.c
