+++ src/kernel/ally.h
@@ -23,6 +23,9 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 extern "C" {
 #endif
 
+    struct attrib_type;
+    extern struct attrib_type at_npcfaction;
+
     typedef struct ally {
         struct ally *next;
         struct faction *faction;
@@ -32,6 +35,15 @@ extern "C" {
     ally * ally_find(ally *al, const struct faction *f);
     ally * ally_add(ally **al_p, struct faction *f);
     void ally_remove(ally **al_p, struct faction *f);
+
+    int AllianceAuto(void);        /* flags that allied factions get automatically */
+    int HelpMask(void);    /* flags restricted to allied factions */
+    int alliedunit(const struct unit *u, const struct faction *f2,
+        int mode);
+    int alliedfaction(const struct plane *pl, const struct faction *f,
+        const struct faction *f2, int mode);
+    int alliedgroup(const struct plane *pl, const struct faction *f,
+        const struct faction *f2, const struct ally *sf, int mode);
 
 #ifdef __cplusplus
 }
