+++ src/reports.c
@@ -26,6 +26,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include "donations.h"
 
 /* kernel includes */
+#include <kernel/ally.h>
 #include <kernel/alliance.h>
 #include <kernel/connection.h>
 #include <kernel/building.h>
